refactor: feedCommentsService 디렉토리 이동
